% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_eliminate.R
\name{batch_eliminate}
\alias{batch_eliminate}
\title{Eliminate alternatives}
\usage{
batch_eliminate(
  round,
  alternative_code,
  continuing_alternatives,
  quietly = TRUE
)
}
\arguments{
\item{round}{An instance of a RC/IR runoff round.}

\item{alternative_code}{Character vector of codes \code{"[A-N]"} for representing
\code{alternatives} in a \code{contest}.}

\item{continuing_alternatives}{A subset of \code{alternative_code} representing
continuing alternatives in a tabulation.}

\item{quietly}{When \code{FALSE} many routines get noisy.}
}
\value{
A character vector of subset of \code{continuing_alternatives}.
}
\description{
Eliminates alternatives by subsetting
  \code{continuing_alternatives}.
}
