% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cvr_contest.R
\name{build_cvr_contest}
\alias{build_cvr_contest}
\title{Build contest from CVR files}
\usage{
build_cvr_contest(CVR)
}
\arguments{
\item{CVR}{A list of dataframes from raw CVR files}
}
\value{
A tibble of cleaned CVR data
}
\description{
Creates the central data frame for a \code{\link{contest}} instance. Typically called by the
  constructor.
}
