% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_cvr.R
\name{classify_cvr}
\alias{classify_cvr}
\title{Classifies CVRs}
\usage{
classify_cvr(path, verbose = TRUE)
}
\arguments{
\item{path}{(character) Path to either a single file CVR or to a directory containing many CVRs (e.g., a folder of JSON or XML files)}

\item{verbose}{(boolean) Controls verbosity of information printed to the user}
}
\value{
(character) One of DELIM, JSON, or XML
}
\description{
Classifies CVRs
}
