% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cvr_alternatives.R
\name{clean_cvr_alternatives}
\alias{clean_cvr_alternatives}
\title{Clean CVR alternatives names}
\usage{
clean_cvr_alternatives(CVR, quietly = TRUE)
}
\value{
Return a \code{\link{contest}} with a clean \code{alternative}.
}
\description{
Creates clean \code{alternative} for contest ranking data.
}
