% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conduct_runoff.R
\name{conduct_runoff}
\alias{conduct_runoff}
\title{Conduct instant runoff}
\usage{
conduct_runoff(
  cntst,
  sequentialU = FALSE,
  max_U = 2,
  sequentialO = FALSE,
  max_O = 1,
  june2018 = FALSE,
  quietly = TRUE
)
}
\arguments{
\item{cntst}{A \code{contest} instance.}

\item{june2018}{If \code{TRUE} duplicates official State tabulations.}

\item{quietly}{When \code{FALSE} many routines get noisy.}
}
\value{
A \code{tabulation} instance.
}
\description{
Conducts a ranked choice instant runoff. Does not resolve ties, as yet.
}
