% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{contest}
\alias{contest}
\title{Create contest object}
\usage{
contest(
  contest_name,
  source = NULL,
  CVR_files = NULL,
  original = FALSE,
  contest_str = NULL,
  date_str = NULL,
  office = NULL,
  jurisdiction = NULL,
  alternatives = NULL,
  ambiguities = NULL,
  creator = NULL,
  quietly = TRUE
)
}
\arguments{
\item{contest_name}{For example "Democratic Primary".
With padding removed, used for first part of the \code{*.rda}
file name for serializing the contest.}

\item{source}{Folder of CVR files.}

\item{CVR_files}{Cast Vote Record file names.}

\item{original}{If \code{original == TRUE}, \code{contest_str} must match
an \code{object} value in \code{contest_keys.txt} file in working directory,
with \code{cvr_url} value
and all \code{cvr_file} values needed to reconstruct the election.}

\item{contest_str}{Must reference an \code{object} value in \code{key_values.txt}.}

\item{date_str}{Encoded as MM/DD/YYYY. Forms the second part of the
\code{*.rda} file name for serializing the contest.}

\item{office}{Encodes the name of the office or referendum at issue.
With padding  removed, forms the third part of the
\code{*.rda} file name for serializing the contest.}

\item{jurisdiction}{Set as appropriate.}

\item{alternatives}{Container for all alternatives in a \code{contest} (char).}

\item{ambiguities}{Container for ambiguities in a \code{contest}.}

\item{creator}{Human creating \code{contest} object.}

\item{quietly}{When \code{FALSE} many routines get noisy.}
}
\value{
An instance of class \code{contest}.
}
\description{
Constructor for instances of S3 class
  \code{contest} implemented as \code{tbl_df} with metadata.
}
