% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_buf.R
\name{download_buf}
\alias{download_buf}
\title{Temporary buffer for downloading file data.}
\usage{
download_buf(url, ext)
}
\arguments{
\item{url}{Universal Resource Locator.}

\item{ext}{File extension.}
}
\value{
A buffer for subsequent reading.
}
\description{
Temporary buffer for downloading file data.
}
\examples{
 download_buf(url = "My favorite webfile", ext= ".txt")

}
