% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_cvr_rankings.R
\name{encode_cvr_rankings}
\alias{encode_cvr_rankings}
\title{Encode cast vote record rankings}
\usage{
encode_cvr_rankings(cntst, cvr_mark_code)
}
\arguments{
\item{cntst}{A \code{contest} instance.}

\item{cvr_mark_code}{The concatenation of \code{alternative_code} and
\code{ambiguity_code}.
The vector of characters to used to encode
each ballot. Alternatives use \code{as.character(LETTERS[1:num_alternatives])}
with U and O for alternatives. \code{names(cvr_mark_code)} corresponds to
entries in the choice fields of \code{cntst}.}
}
\value{
A \code{rank_str})
  derived from ranking data for each record.
}
\description{
Encodes rankings on each ranked choice cvr as a single
  string \code{rank_str} with a character [A-Z] for each
  ranking, designating a specific \code{alternative_code}
  or \code{ambiguity_code}.
}
