% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glean_cvr_alternatives.R
\name{glean_cvr_names}
\alias{glean_cvr_names}
\title{Glean CVR \code{alternatives}}
\usage{
glean_cvr_names(CVR)
}
\arguments{
\item{CVR}{}
}
\value{

}
\description{
Assumes the data confined to columns \code{4:(length(CVR) - 1)}.
  Will be modified or extended for other jurisdictions.
}
