% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{header_processor}
\alias{header_processor}
\title{Process delimited files with unconventional headers. They are
generated by one of the voting machine vendors, and include
the name of the election in the top left, then a blank line, then
three rows of information on each column. Sometimes the order of the rows varies
but they always include the contest, the name of the candidate, and their party.
In addition, the first m columns are metadata columns related to the CVR. This
function processes these files into a more regular delimited format, with a singular
row for the header. There is also an edge case in some counties (e.g., San Diego, CA 2020)
wherein the header is inexplicably repeated partway through the file. The chunk that
generates \code{bad_rows} removes this extra information.}
\usage{
header_processor(path, n = Inf)
}
\arguments{
\item{path}{The path to the file}

\item{n}{The number of rows to read. This is used to generate the raw contests
quickly, without reading the entire file.}
}
\value{
A tibble formatted with conventional headers
}
\description{
Process delimited files with unconventional headers. They are
generated by one of the voting machine vendors, and include
the name of the election in the top left, then a blank line, then
three rows of information on each column. Sometimes the order of the rows varies
but they always include the contest, the name of the candidate, and their party.
In addition, the first m columns are metadata columns related to the CVR. This
function processes these files into a more regular delimited format, with a singular
row for the header. There is also an edge case in some counties (e.g., San Diego, CA 2020)
wherein the header is inexplicably repeated partway through the file. The chunk that
generates \code{bad_rows} removes this extra information.
}
