% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_header}
\alias{is_header}
\title{Check if delimited file has an unconventional header}
\usage{
is_header(path)
}
\arguments{
\item{path}{the path to the delimited file}
}
\value{
boolean
}
\description{
Check if delimited file has an unconventional header
}
