% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_contest.R
\name{new_contest}
\alias{new_contest}
\title{Instantiate \code{\link{contest}} object carefully}
\usage{
new_contest(
  x,
  CVRdata,
  contest_name,
  source,
  CVR_files,
  original,
  contest_str,
  date_str,
  office,
  alternatives,
  ambiguities = c("undervote", "overvote"),
  jurisdiction,
  creator,
  num_ranks,
  num_records,
  num_alternatives,
  first_rank_col
)
}
\arguments{
\item{contest_name}{For example "Democratic Primary".
With padding removed, used for first part of the \code{*.rda}
file name for serializing the contest.}

\item{source}{Folder of CVR files.}

\item{CVR_files}{Cast Vote Record file names.}

\item{original}{If \code{original == TRUE}, \code{contest_str} must match
an \code{object} value in \code{contest_keys.txt} file in working directory,
with \code{cvr_url} value
and all \code{cvr_file} values needed to reconstruct the election.}

\item{contest_str}{Must reference an \code{object} value in \code{key_values.txt}.}

\item{date_str}{Encoded as MM/DD/YYYY. Forms the second part of the
\code{*.rda} file name for serializing the contest.}

\item{office}{Encodes the name of the office or referendum at issue.
With padding  removed, forms the third part of the
\code{*.rda} file name for serializing the contest.}

\item{alternatives}{Container for all alternatives in a \code{contest} (char).}

\item{ambiguities}{Container for ambiguities in a \code{contest}.}

\item{jurisdiction}{Set as appropriate.}

\item{creator}{Human creating \code{contest} object.}
}
\value{

}
\description{
Instantiate \code{\link{contest}} object carefully
}
\examples{

}
