% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{rcvr parameters}
\arguments{
\item{alternative}{Candidate or alternative for a \code{\link{contest}} (char).}

\item{alternatives}{Container for all alternatives in a \code{contest} (char).}

\item{alternative_code}{Character vector of codes \code{"[A-N]"} for representing
\code{alternatives} in a \code{contest}.}

\item{ambiguity}{The string found on the CVR representing an unresolved ranking.}

\item{ambiguities}{Container for ambiguities in a \code{contest}.}

\item{contest_name}{For example "Democratic Primary".
With padding removed, used for first part of the \code{*.rda}
file name for serializing the contest.}

\item{contest_str}{Must reference an \code{object} value in \code{key_values.txt}.}

\item{continuing_alternatives}{A subset of \code{alternative_code} representing
continuing alternatives in a tabulation.}

\item{cntst}{A \code{contest} instance.}

\item{creator}{Human creating \code{contest} object.}

\item{CVR}{A list of dataframes with cast vote records. Usually raw data saved
used by the constructor and saved as metadata for a \code{contest} instance.}

\item{CVR_files}{Cast Vote Record file names.}

\item{date_str}{Encoded as MM/DD/YYYY. Forms the second part of the
\code{*.rda} file name for serializing the contest.}

\item{june2018}{If \code{TRUE} duplicates official State tabulations.}

\item{jurisdiction}{Set as appropriate.}

\item{source}{Folder of CVR files.}

\item{office}{Encodes the name of the office or referendum at issue.
With padding  removed, forms the third part of the
\code{*.rda} file name for serializing the contest.}

\item{original}{If \code{original == TRUE}, \code{contest_str} must match
an \code{object} value in \code{contest_keys.txt} file in working directory,
with \code{cvr_url} value
and all \code{cvr_file} values needed to reconstruct the election.}

\item{path}{Some folder, somewhere. Typically defaults to working directory.}

\item{quietly}{When \code{FALSE} many routines get noisy.}

\item{round}{An instance of a RC/IR runoff round.}

\item{url}{Universal Resource Locator.}
}
\description{
The rcvr package provides functions for reading, analyzing, and visualizing
  ballot data from Ranked Choice Voting elections.

  When \code{\link{rcvr}} is installed, missing *url* arguments default to the
  *extdata* subfolder of the installed *rcvr* package folder on the local machine
  (*system.file("extdata", package = "rcvr").
  If space is an issue, the utility *dump_default_files()* is provided for your
  cobvenience.

  For ease of of development, second copy of default data files are installed the
  *extdata* subfolder of the *rcvr* package development directory
  (*system.file("extdata", package = "rcvr*).
}
