% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_contest_keys.R
\name{read_contest_keys}
\alias{read_contest_keys}
\title{Read contest keys}
\usage{
read_contest_keys(contest_str, path = getwd())
}
\arguments{
\item{contest_str}{Must reference an \code{object} value in \code{key_values.txt}.}

\item{path}{Some folder, somewhere. Typically defaults to working directory.}
}
\value{
Names for an election
}
\description{
Get names and such for loading election data.
}
\examples{
read_key_values(contest_str = "d2_June", path = "My data folder")

}
