% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_delim.R
\name{read_delim_multi_cvr}
\alias{read_delim_multi_cvr}
\title{Read a directory of delimited CVR files}
\usage{
read_delim_multi_cvr(dir)
}
\arguments{
\item{dir}{Path to a directory of CSV or Excel CVRs.}
}
\value{
A wide tibble, all files row-bound.
}
\description{
Read a directory of delimited CVR files
}
