% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read_json_cvr}
\alias{read_json_cvr}
\title{Read a directory of Dominion JSON CVRs}
\usage{
read_json_cvr(path)
}
\arguments{
\item{path}{Path to the directory holding \code{CvrExport*.json} and the
Dominion manifests.}
}
\value{
A pairs frame.
}
\description{
Read a directory of Dominion JSON CVRs
}
