% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ME_contest.R
\name{read_ME_contest}
\alias{read_ME_contest}
\title{Read Maine contest files}
\usage{
read_ME_contest(source, CVR_files, original = FALSE, quietly = TRUE)
}
\arguments{
\item{source}{Folder of CVR files.}

\item{CVR_files}{Cast Vote Record file names.}

\item{original}{If \code{original == TRUE}, \code{contest_str} must match
an \code{object} value in \code{contest_keys.txt} file in working directory,
with \code{cvr_url} value
and all \code{cvr_file} values needed to reconstruct the election.}

\item{quietly}{When \code{FALSE} many routines get noisy.}
}
\value{
A list of dataframes.
}
\description{
Used to read Cast Vote Record files (\code{*.xlsx}) for a State of Maine
  Ranked Choice Instant Runoff election.
}
