% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ME_CVR_file.R
\name{read_ME_CVR_file}
\alias{read_ME_CVR_file}
\title{Read Maine CVR file}
\usage{
read_ME_CVR_file(url, original = FALSE, quietly = TRUE)
}
\arguments{
\item{url}{Universal Resource Locator.}

\item{original}{If \code{original == TRUE}, \code{contest_str} must match
an \code{object} value in \code{contest_keys.txt} file in working directory,
with \code{cvr_url} value
and all \code{cvr_file} values needed to reconstruct the election.}

\item{quietly}{When \code{FALSE} many routines get noisy.}
}
\value{
A tibble of raw CVR data, Maine style.
}
\description{
Read Maine CVR file
}
