% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_names.R
\name{replace_names}
\alias{replace_names}
\title{Replace names}
\usage{
replace_names(x, new_names)
}
\arguments{
\item{x}{A vector of values (char).}

\item{new_names}{A vector of new names for the values (char).}
}
\value{
A vector with new names.
}
\description{
Puts new names for a column of choices in CVR ballot dataframe
  Called by code{clean_cvr_alternatives()} using code{lapply()} to
  create cleaner CVR table for a \code{contest} instance..
}
