% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_rankings.R
\name{standardize_rankings}
\alias{standardize_rankings}
\title{Standardize rankings}
\usage{
standardize_rankings(x, ext = "")
}
\arguments{
\item{x}{Dataframe containing Maine cast vote records.}
}
\value{
A tibble.
}
\description{
Standardizes \code{rankings}. Assumes Maine cast Vote record style.
}
