% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_cvr_rank.R
\name{tabulate_cvr_rank}
\alias{tabulate_cvr_rank}
\title{Tabulate CVR rank string}
\usage{
tabulate_cvr_rank(
  rank_str,
  continuing_alternatives,
  sequentialO = FALSE,
  max_O = 1,
  sequentialU = FALSE,
  max_U = 2,
  june2018 = FALSE
)
}
\arguments{
\item{sequentialO}{If \code{TRUE}, \code{max_O} sequential overvotes \code{exhaust}
the cvr by undervote, \code{status_code = "exo"}.}

\item{max_O}{The number of overvotes triggering a potential overvote,
according to the value of \code{sequentialO}.}

\item{sequentialU}{If \code{TRUE}, \code{max_u} sequential undervotes \code{exhaust}
the cvr by undervote, \code{status_code = "exu"}.}

\item{max_U}{The number of undervotes triggering a potential undervote,
according to the value of \code{sequentialO}.}
}
\value{
A value from \code{status_code}
}
\description{
Implements rules for the disposition of a cast vote record \code{rank_str},
  including the disposition of ambiguities such as overvotes and undervotes.
}
