% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_round.R
\name{tabulate_round}
\alias{tabulate_round}
\title{Tabulate instant runoff round}
\usage{
tabulate_round(cntst, result, cvr_status_code, round_number)
}
\arguments{
\item{result}{A \code{list} of results from \code{tabulate_cvr_rank}.}

\item{cvr_status_code}{Codes used encode disposition of a cvr during an instant runoff round.}

\item{round_number}{The runoff round number.}
}
\value{
A \code{tabulation} instance.
}
\description{
Tabulates the \code{result} of an instant runoff round for a ranked choice \code{contest}.
}
